fix: 컨트랙트 코인 양 조회시 DB에 업데이트 하도록 수정

+++ src/main/java/com/baekho/bridgenet/domain/bridge/dto/response/ChainGetDetailDTO.java
@@ -3,12 +3,14 @@ package com.baekho.bridgenet.domain.bridge.dto.response;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+import java.math.BigInteger;
+
 @Getter
 @AllArgsConstructor
 public class ChainGetDetailDTO {
     private Long chainId;
     private String chainName;
     private String smartContractAddress;
-    private Long smartContractValue;
+    private BigInteger smartContractValue;
     private String unit;
 }
